Fix post-rebase issues
This commit adds some fixes that were introduced during the rebasing
process.
