+++ src/providers/trusted_service/mod.rs
@@ -5,20 +5,19 @@
 //! This provider is backed by a crypto Trusted Service deployed in TrustZone
 use super::mbed_crypto::key_management as mbed_crypto_key_management;
 use crate::authenticators::ApplicationName;
-use crate::key_info_managers::{self, KeyTriple, ManageKeyInfo};
+use crate::key_info_managers::{KeyTriple, ManageKeyInfo};
 use crate::providers::Provide;
 use context::Context;
 use derivative::Derivative;
 use log::{error, trace};
 use parsec_interface::operations::list_providers::ProviderInfo;
 use parsec_interface::operations::{
-    list_keys, psa_destroy_key, psa_export_public_key, psa_generate_key, psa_import_key,
-    psa_sign_hash, psa_verify_hash,
+    list_clients, list_keys, psa_destroy_key, psa_export_public_key, psa_generate_key,
+    psa_import_key, psa_sign_hash, psa_verify_hash,
 };
 use parsec_interface::requests::{Opcode, ProviderID, ResponseStatus, Result};
 use psa_crypto::types::key;
 use std::collections::HashSet;
-use std::ops::Deref;
 use std::sync::{
     atomic::{AtomicU32, Ordering},
     Arc, RwLock,
@@ -143,15 +142,27 @@ impl Provide for Provider {
     ) -> Result<list_keys::Result> {
         let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
         Ok(list_keys::Result {
-            keys: key_info_managers::list_keys(
-                store_handle.deref(),
-                &app_name,
-                ProviderID::TrustedService,
-            )
-            .map_err(|e| {
-                format_error!("Error occurred when fetching key information", e);
-                ResponseStatus::KeyInfoManagerError
-            })?,
+            keys: store_handle
+                .list_keys(&app_name, ProviderID::TrustedService)
+                .map_err(|e| {
+                    format_error!("Error occurred when fetching key information", e);
+                    ResponseStatus::KeyInfoManagerError
+                })?,
+        })
+    }
+
+    fn list_clients(&self, _op: list_clients::Operation) -> Result<list_clients::Result> {
+        let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
+        Ok(list_clients::Result {
+            clients: store_handle
+                .list_clients(ProviderID::TrustedService)
+                .map_err(|e| {
+                    format_error!("Error occurred when fetching key information", e);
+                    ResponseStatus::KeyInfoManagerError
+                })?
+                .into_iter()
+                .map(|app_name| app_name.to_string())
+                .collect(),
         })
     }
